% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canope_call_cnvs.R
\name{calc_copy_number}
\alias{calc_copy_number}
\title{Estimate Most-Likely Copy Number for Each CNV (negative-binomial states C0-C6)}
\usage{
calc_copy_number(data, cnvs, homdel_mean)
}
\description{
Estimate Most-Likely Copy Number for Each CNV (negative-binomial states C0-C6)
}
