% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canope_utils.R
\name{parse_canope_targets}
\alias{parse_canope_targets}
\title{Parse a CANOPE TARGETS string into target ID range}
\usage{
parse_canope_targets(targets)
}
\arguments{
\item{targets}{Character string of the form \code{"42..47"}.}
}
\value{
Integer vector \code{c(start_target, end_target)}.
}
\description{
Parse a CANOPE TARGETS string into target ID range
}
