% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canope_utils.R
\name{stop_if_not_file}
\alias{stop_if_not_file}
\title{Stop if a file does not exist}
\usage{
stop_if_not_file(path, msg)
}
\arguments{
\item{path}{File path.}

\item{msg}{Error message.}
}
\description{
Stop if a file does not exist
}
